refactor: rename file, change tags html and name classes - rename component file - rename tags to HTML5 standards - rename classesName to method BEM standards - move dir Login to Pages
